+++ coverage.txt
@@ -1,5 +1,36 @@
-github.com/sbilibin2017/go-gophermart/cmd/accrual/app/command.go:25:	NewCommand	100.0%
-github.com/sbilibin2017/go-gophermart/cmd/accrual/app/config.go:8:	GetRunAddress	100.0%
-github.com/sbilibin2017/go-gophermart/cmd/accrual/app/config.go:12:	GetDatabaseURI	100.0%
-github.com/sbilibin2017/go-gophermart/cmd/accrual/app/run.go:5:		Run		100.0%
-github.com/sbilibin2017/go-gophermart/pkg/cli/run.go:5:			Run		100.0%
+github.com/sbilibin2017/go-gophermart/cmd/accrual/app/command.go:25:			NewCommand				100.0%
+github.com/sbilibin2017/go-gophermart/cmd/accrual/app/config.go:8:			GetRunAddress				100.0%
+github.com/sbilibin2017/go-gophermart/cmd/accrual/app/config.go:12:			GetDatabaseURI				100.0%
+github.com/sbilibin2017/go-gophermart/cmd/accrual/app/run.go:5:				Run					100.0%
+github.com/sbilibin2017/go-gophermart/cmd/accrual/app/server.go:19:			NewServer				0.0%
+github.com/sbilibin2017/go-gophermart/internal/handlers/order_register.go:19:		OrderRegisterHandler			100.0%
+github.com/sbilibin2017/go-gophermart/internal/handlers/order_register.go:40:		handleOrderRegisterError		100.0%
+github.com/sbilibin2017/go-gophermart/internal/json/request_decoder.go:12:		NewRequestDecoder			100.0%
+github.com/sbilibin2017/go-gophermart/internal/json/request_decoder.go:16:		Decode					100.0%
+github.com/sbilibin2017/go-gophermart/internal/log/log.go:19:				Init					100.0%
+github.com/sbilibin2017/go-gophermart/internal/log/log.go:37:				Info					100.0%
+github.com/sbilibin2017/go-gophermart/internal/log/log.go:45:				Error					100.0%
+github.com/sbilibin2017/go-gophermart/internal/log/log.go:53:				Sync					100.0%
+github.com/sbilibin2017/go-gophermart/internal/middlewares/gzip.go:10:			GzipMiddleware				100.0%
+github.com/sbilibin2017/go-gophermart/internal/middlewares/gzip.go:41:			Write					100.0%
+github.com/sbilibin2017/go-gophermart/internal/middlewares/logging.go:10:		LoggingMiddleware			100.0%
+github.com/sbilibin2017/go-gophermart/internal/middlewares/logging.go:34:		WriteHeader				100.0%
+github.com/sbilibin2017/go-gophermart/internal/middlewares/logging.go:39:		Write					100.0%
+github.com/sbilibin2017/go-gophermart/internal/repositories/order_exists.go:17:		NewOrderExistRepository			0.0%
+github.com/sbilibin2017/go-gophermart/internal/repositories/order_exists.go:22:		Exists					0.0%
+github.com/sbilibin2017/go-gophermart/internal/repositories/order_save.go:16:		NewOrderSaveRepository			0.0%
+github.com/sbilibin2017/go-gophermart/internal/repositories/order_save.go:20:		Save					0.0%
+github.com/sbilibin2017/go-gophermart/internal/repositories/reward_filter.go:16:	NewRewardFilterRepository		0.0%
+github.com/sbilibin2017/go-gophermart/internal/repositories/reward_filter.go:20:	getRewardFilterQueryArgs		0.0%
+github.com/sbilibin2017/go-gophermart/internal/repositories/reward_filter.go:28:	Filter					0.0%
+github.com/sbilibin2017/go-gophermart/internal/routers/order_register.go:9:		RegisterOrderRegisterRoute		100.0%
+github.com/sbilibin2017/go-gophermart/internal/services/order_register.go:40:		NewOrderRegisterService			0.0%
+github.com/sbilibin2017/go-gophermart/internal/services/order_register.go:61:		Register				0.0%
+github.com/sbilibin2017/go-gophermart/internal/storage/db.go:12:			NewDB					100.0%
+github.com/sbilibin2017/go-gophermart/internal/storage/tx.go:15:			NewTx					100.0%
+github.com/sbilibin2017/go-gophermart/internal/storage/tx.go:19:			Do					100.0%
+github.com/sbilibin2017/go-gophermart/internal/usecases/order_register.go:23:		NewOrderRegisterUsecase			100.0%
+github.com/sbilibin2017/go-gophermart/internal/usecases/order_register.go:33:		Execute					100.0%
+github.com/sbilibin2017/go-gophermart/internal/validators/louna.go:11:			ValidateNumberWithLouna			100.0%
+github.com/sbilibin2017/go-gophermart/internal/validators/louna.go:41:			RegisterLounaValidator			100.0%
+github.com/sbilibin2017/go-gophermart/pkg/cli/run.go:5:					Run					100.0%
