+++ pkg/db/db.go
@@ -1,4 +1,4 @@
-package storage
+package db
 
 import (
 	_ "github.com/jackc/pgx/v5/stdlib"
