+++ pkg/db/db_test.go
@@ -1,4 +1,4 @@
-package storage
+package db
 
 import (
 	"testing"
