+++ pkg/db/setup_test.go
@@ -1,4 +1,4 @@
-package storage
+package db
 
 import (
 	"context"
