+++ pkg/log/log_test.go
