+++ pkg/srv/run_test.go
@@ -1,4 +1,4 @@
-package server
+package srv
 
 import (
 	"context"
